services/actor/requirements.txt
pyyaml>=6.0
requests>=2.31.0
flask>=2.3.0
elevenlabs>=1.0.0
